Refactor for abstraction
- Amount struct with conversion methods and hidden internals
- Remove all pointers and dereferences. Unneeded. This should be faster (stay on stack)
- Remove calculator file and move all calculations to new Amount object

+++ formatter_test.go
@@ -11,7 +11,7 @@ func TestFormatter_Format(t *testing.T) {
 		thousand string
 		grapheme string
 		template string
-		amount   int
+		amount   int64
 		expected string
 	}{
 		{2, ".", ",", "$", "1 $", 0, "0.00 $"},
